commonsys: services: upgrade IDC 2.0 for backward compatible
The IDC libs in vendor and system images are not compatible due
to diverged changes. E.g., in legacy vendor kGetDisplayTileCount
op code is 50, while in new system images op code 50 stands for
kDummyOpCode. This breaks the backward compatibility. So upgrade
IDC lib in new system to fix the backward compatible issues for
vendor freeze targets.

Change-Id: Ia7552d2f06db04db149fd7fe3bf875c1406421c8
Signed-off-by: GhostMaster69-dev <rathore6375@gmail.com>